UserRegSystem/requirements.txt
flask
mysql-connector-python
matplotlib
